# docs/openapi/pecr.v1.yaml
openapi: 3.1.0
info:
  title: PECR API
  version: 1.0.0
  description: |
    Versioned API contract for PECR v1.
servers:
  - url: http://127.0.0.1:8080
    description: Gateway
  - url: http://127.0.0.1:8081
    description: Controller
paths:
...
  /v1/finalize:
    post:
      summary: Finalize response with claim map gate
      operationId: finalize
      security:
        - PecrSessionToken: []
          PecrLocalPrincipal: []
        - PecrSessionToken: []
          BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XPecrSessionToken"
        - $ref: "#/components/parameters/XPecrRequestId"
        - $ref: "#/components/parameters/XPecrTraceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [session_id, response_text, claim_map]
              properties:
                session_id:
                  type: string
                response_text:
                  type: string
                claim_map:
                  $ref: "#/components/schemas/ClaimMap"
      responses:
        "200":
          description: Finalized
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [terminal_mode, trace_id, claim_map, response_text]
                properties:
                  terminal_mode:
                    $ref: "#/components/schemas/TerminalMode"
                  trace_id:
                    type: string
                  claim_map:
                    $ref: "#/components/schemas/ClaimMap"
                  response_text:
                    type: string
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Denied by policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Session already finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Policy engine timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Source unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    PecrLocalPrincipal:
      type: apiKey
      in: header
      name: x-pecr-principal-id
      description: Local mode principal identity header (`PECR_AUTH_MODE=local`).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC bearer token (`PECR_AUTH_MODE=oidc`).
    PecrSessionToken:
      type: apiKey
      in: header
      name: x-pecr-session-token
      description: Opaque session capability token returned from `POST /v1/sessions`.
  parameters:
    XPecrSessionToken:
      name: x-pecr-session-token
      in: header
      required: true
      schema:
        type: string
      description: Required for `/v1/operators/{op_name}` and `/v1/finalize`.
    XPecrRequestId:
      name: x-pecr-request-id
      in: header
      required: false
      schema:
        type: string
      description: Optional caller-provided id used for log correlation.
    XPecrTraceId:
      name: x-pecr-trace-id
      in: header
      required: false
      schema:
        type: string
      description: Optional trace id propagated across controller and gateway.
  schemas:
...
    TerminalMode:
      type: string
      enum:
        - SUPPORTED
        - INSUFFICIENT_EVIDENCE
        - INSUFFICIENT_PERMISSION
        - SOURCE_UNAVAILABLE
    ClientResponseKind:
      type: string
      description: |
        Optional client UX hint for `/v1/run` and error responses.
        Absent means the response is a normal supported answer without special handling.
      enum:
        - blocked
        - ambiguous
        - source_down
        - partial_answer
...
    ClaimEvidenceSnippet:
      type: object
      additionalProperties: false
      required:
        - evidence_unit_id
        - location
        - snippet
      properties:
        evidence_unit_id:
          type: string
          pattern: "^[a-f0-9]{64}$"
        location:
          type: string
          minLength: 1
        snippet:
          type: string
          minLength: 1
    Claim:
      type: object
      additionalProperties: false
      required:
        - claim_id
        - claim_text
        - status
        - evidence_unit_ids
      properties:
        claim_id:
          type: string
          pattern: "^[a-f0-9]{64}$"
        claim_text:
          type: string
          minLength: 1
        status:
          type: string
          enum:
            - SUPPORTED
            - ASSUMPTION
            - UNKNOWN
        evidence_unit_ids:
          type: array
          items:
            type: string
            pattern: "^[a-f0-9]{64}$"
        evidence_snippets:
          type: array
          items:
            $ref: "#/components/schemas/ClaimEvidenceSnippet"
...
    ClarificationPrompt:
      type: object
      additionalProperties: false
      required:
        - question
      properties:
        question:
          type: string
          minLength: 1
          description: Short narrowing question the client can render directly.
        options:
          type: array
          description: Optional suggested follow-up choices for quick-pick UI.
          items:
            type: string
            minLength: 1
    ClaimMap:
      type: object
      additionalProperties: false
      required:
        - claim_map_id
        - terminal_mode
        - claims
        - coverage_threshold
        - coverage_observed
      properties:
        claim_map_id:
          type: string
        terminal_mode:
          $ref: "#/components/schemas/TerminalMode"
        claims:
          type: array
          items:
            $ref: "#/components/schemas/Claim"
        coverage_threshold:
          type: number
          minimum: 0
          maximum: 1
        coverage_observed:
          type: number
          minimum: 0
          maximum: 1
        clarification_prompt:
          $ref: "#/components/schemas/ClarificationPrompt"
        notes:
          type: string
          description: Additional client-displayable guidance, for example partial-answer context.
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - terminal_mode_hint
        - retryable
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        terminal_mode_hint:
          $ref: "#/components/schemas/TerminalMode"
        retryable:
          type: boolean
        response_kind:
          $ref: "#/components/schemas/ClientResponseKind"
        what_failed:
          type: string
          minLength: 1
          description: Concrete explanation of the blocked or unavailable step.
        safe_alternative:
          type: string
          minLength: 1
          description: Suggested safe next step for the caller or user.
        detail:
          oneOf:
            - type: object
            - type: array
            - type: string
            - type: number
            - type: integer
            - type: boolean
            - type: null